requirements.txt
fastapi
uvicorn
sqlalchemy
passlib[bcrypt]
python-jose
pydantic[email]
psycopg2-binary
python-dotenv
openai>=1.12.0
six==1.16.0
email-validator>=2.0.0

